# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ClawOPT backend",
    "version": "1.0.0",
    "description": "Generated from the backend route registry (npm run openapi:generate). Do not edit by hand."
  },
...
  "paths": {
...
    "/api/providers/audit": {
      "get": {
        "operationId": "getApiProvidersAudit",
        "tags": [
          "control/models"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/providers/audit",
        "x-clawopt-order": 114
      }
    },
    "/api/providers/{id}/catalog": {
      "get": {
        "operationId": "getApiProvidersIdCatalog",
        "tags": [
          "control/models"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": false,
        "x-clawopt-express-path": "/api/providers/:id/catalog",
        "x-clawopt-order": 115
      }
    },
    "/api/providers/{id}/catalog/refresh": {
      "post": {
        "operationId": "postApiProvidersIdCatalogRefresh",
        "tags": [
          "control/models"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/providers/:id/catalog/refresh",
        "x-clawopt-order": 116
      }
    },
    "/api/providers/{id}/catalog/restore": {
      "post": {
        "operationId": "postApiProvidersIdCatalogRestore",
        "tags": [
          "control/models"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/providers/:id/catalog/restore",
        "x-clawopt-order": 117
      }
    },
    "/api/providers/{id}/visibility": {
      "put": {
        "operationId": "putApiProvidersIdVisibility",
        "tags": [
          "control/models"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/providers/:id/visibility",
        "x-clawopt-order": 118
      }
    },
    "/api/providers/{id}/context-lengths": {
      "put": {
        "operationId": "putApiProvidersIdContextLengths",
        "tags": [
          "control/models"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/providers/:id/context-lengths",
        "x-clawopt-order": 119
      }
    },
    "/api/models/aliases": {
      "get": {
        "operationId": "getApiModelsAliases",
        "tags": [
          "control/models"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": false,
        "x-clawopt-express-path": "/api/models/aliases",
        "x-clawopt-order": 120
      }
    },
    "/api/models/aliases/{alias}": {
      "put": {
        "operationId": "putApiModelsAliasesAlias",
        "tags": [
          "control/models"
        ],
        "parameters": [
          {
            "name": "alias",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/models/aliases/:alias",
        "x-clawopt-order": 121
      },
      "delete": {
        "operationId": "deleteApiModelsAliasesAlias",
        "tags": [
          "control/models"
        ],
        "parameters": [
          {
            "name": "alias",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/models/aliases/:alias",
        "x-clawopt-order": 122
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "clawopt_session",
        "description": "httpOnly session cookie issued by POST /api/auth/login (web UI)."
      },
      "authTokenHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-ClawOPT-Auth-Token",
        "description": "Session token for CLI and scripts."
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Same session token sent as Authorization: Bearer."
      }
    }
  }
}